float::sin - improve the Taylor series and test
Improve the Taylor series by using bigints and allow to run longer
sequences, and also bail early, if the series converges. Also add a
test.  Notice that we still had a precision problem with large numbers
that rely on modulo pi, where pi is not accurate enough for repetitive
summation.