Implement efficient powi()
This algorithm is similar to binary conversion. Each bit in 'n'
represents a power-of-two number, like 1,2,4,8 ... We know how to
generate numbers to the power of an even number by squaring the number
log2 times. So, we just multiply all of the numbers together to get the
result. This is like converting a binary number to integer except that
instead of adding we multiply the values.