Optimize Div
Optimize Div for the case where the divisor is shorter than the
dividend.  When this happens we know that there are zeros at the lower
part of the word because we shifted the divisor left. Knowing that we
have zeros allows us to subtract shorter values.

This speeds up a few benchmarks by 10%-15%.