Optimize karatsuba for inputs of different sizes.
Optimize karatsuba for inputs of different sizes. Before this commit we
split the words in half, where half was defined by the smaller part.
This was not ideal for inputs such as [2] x [100] where the recursion
just pealed one word. Instead we split the words in half, and allow one
of the halves to be of zero-length.